+++ backend/canisters/cycles_dispenser/impl/src/lifecycle/pre_upgrade.rs
@@ -1,8 +1,8 @@
-use crate::lifecycle::UPGRADE_BUFFER_SIZE;
+use crate::memory::get_upgrades_memory;
 use crate::take_state;
 use canister_tracing_macros::trace;
 use ic_cdk_macros::pre_upgrade;
-use stable_memory::serialize_to_stable_memory;
+use stable_memory::get_writer;
 use tracing::info;
 
 #[pre_upgrade]
@@ -16,5 +16,8 @@ fn pre_upgrade() {
 
     let stable_state = (state.data, logs, traces);
 
-    serialize_to_stable_memory(stable_state, UPGRADE_BUFFER_SIZE).unwrap();
+    let mut memory = get_upgrades_memory();
+    let writer = get_writer(&mut memory);
+
+    serializer::serialize(stable_state, writer).unwrap();
 }