+++ backend/canisters/proposal_validation/api/src/infinity_swap/mod.rs
@@ -1,2 +1,4 @@
+pub mod burn;
 pub mod mint;
 pub mod receive_tokens;
+pub mod refund_transfer;