+++ backend/canisters/storage_index/impl/src/lifecycle/mod.rs
@@ -11,8 +11,6 @@ mod init;
 mod post_upgrade;
 mod pre_upgrade;
 
-const BUFFER_SIZE: usize = 4 * 1024 * 1024; // 4MB
-
 fn init_env() -> Box<CanisterEnv> {
     ic_cdk_timers::set_timer(Duration::ZERO, reseed_rng);
     Box::default()