+++ backend/libraries/chat_events/src/lib.rs
@@ -3,10 +3,12 @@ mod chat_events;
 mod chat_events_list;
 mod expiring_events;
 mod last_updated_timestamps;
+mod message_content_internal;
 
 pub use crate::chat_event_internal::*;
 pub use crate::chat_events::*;
 pub use crate::chat_events_list::*;
+pub use crate::message_content_internal::*;
 
 fn incr(counter: &mut u64) {
     *counter = counter.saturating_add(1);