+++ backend/canisters/notifications_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Use dynamic buffer size when reading from stable memory ([#4683](https://github.com/open-chat-labs/open-chat/pull/4683))
+
 ## [[2.0.899](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.899-notifications_index)] - 2023-10-19
 
 ### Changed